Connect to default database to find all databases
The psql call for finding all databases (which happens when DB_NAME is
ALL or SPLIT_DB is FALSE) fails if DB_USER is not an existing database.
This change stops using DB_USER for the -d option.

If you run psql without specifying -d/--dbname, it uses "postgres" by
default, going so far as to raise an error if it is missing:

    / # psql -U postgres
    psql: error: FATAL:  database "postgres" does not exist

I think it makes sense to use the same assumption for the psql call
mentioned above.